Rename package 'com.plattysoft.leonids' to 'androidx.particles'
The library's group has been renamed to 'one.atpc.thomorl' and the
package change has been reflected in the library's AndroidManifest.
