+++ LeonidsExamples/src/main/java/com/plattysoft/leonids/examples/ConfettiExampleActivity.java
@@ -1,6 +1,6 @@
 package com.plattysoft.leonids.examples;
 
-import com.plattysoft.leonids.ParticleSystem;
+import androidx.particles.ParticleSystem;
 
 import android.app.Activity;
 import android.os.Bundle;
