+++ LeonidsLib/src/main/java/androidx/particles/initializers/SpeeddByComponentsInitializer.java
@@ -1,8 +1,8 @@
-package com.plattysoft.leonids.initializers;
+package androidx.particles.initializers;
 
 import java.util.Random;
 
-import com.plattysoft.leonids.Particle;
+import androidx.particles.Particle;
 
 public class SpeeddByComponentsInitializer implements ParticleInitializer {
